+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Add filter icon to filters active info node
 - Fix missing context flag for show export button
+- Implement a sub tag regex (based on work by[calw20](https://github.com/calw20))
 
 ## v0.0.196 - 2021-01-21